refactor: Split catalog/store/advanced.rs into separate modules
Split the 712-line advanced.rs file into focused modules organized by
SQL object type. This improves maintainability and follows the Single
Responsibility Principle.

Module organization:
- types.rs (80 lines) - User-defined type management
- domains.rs (107 lines) - Domain management
- sequences.rs (187 lines) - Sequence management
- collations.rs (32 lines) - Collation management
- charsets.rs (30 lines) - Character set management
- translations.rs (34 lines) - Translation management
- views.rs (115 lines) - View management + dependency tracking
- triggers.rs (52 lines) - Trigger management
- assertions.rs (46 lines) - Assertion management
- routines.rs (74 lines) - Function and procedure management
- mod.rs (40 lines) - Module organization and documentation

Benefits:
- Each module handles a single SQL object type
- Easier to navigate and find type-specific code
- Better organized for testing individual object types
- Enables parallel development on different object types
- Clearer separation of view/trigger dependency logic

All catalog tests pass with no regressions.

Closes #1146
